minor modifications

+++ dbmanagement/manager/category_manager.py
@@ -7,7 +7,7 @@ class CategoryManager:
     def create_tables(self):
         """Create the Category table"""
         db.query(""" CREATE TABLE IF NOT EXISTS Category (
-                          id BIGINT UNSIGNED PRIMARY KEY AUTO_INCREMENT,
+                          id INT UNSIGNED PRIMARY KEY AUTO_INCREMENT,
                           category_name VARCHAR(255) UNIQUE);
                       """)
